+++ app/src/main/java/com/dod/whateat/data/SeqData.kt
@@ -1,5 +1,5 @@
 package com.dod.whateat.data
 
 interface SeqData {
-    open val seq: Int
+    open val seq: Long
 }
